enhance custom tailwind plugin for hover effects
The custom plugin has the 'has-hover' variant to enable hover effects only on devices that support hover. The idea was to prevent hover effects on devices with touch screens which usually don't support hover. However, I discovered that Android devices support keyboard and mouse by default (this includes mobile phones), for this reason the original implementation didn't prevent hover on such devices.

The query was modified to trigger the hover effects only when the device support hover effects and uses a mouse pointer to navigate, this should prevent hover events on Android device when the touch screen is used.